% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_offset.R
\name{add_offset}
\alias{add_offset}
\title{Add an offset, in units of the datum, to a timestamp}
\usage{
add_offset(datum, offsets)
}
\arguments{
\item{datum}{CFdatum. The datum that defines the unit of the offsets and the
origin to add the offsets to.}

\item{offsets}{numeric. Vector of offsets to add to the datum.}
}
\value{
A tibble with columns year, month, day, hour, minute, second, offset.
}
\description{
Add an offset, in units of the datum, to a timestamp
}
\examples{
datum <- CFdatum("days since 1850-01-01", "gregorian")
add_offset(datum, 0:364)
}
