% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFtime.R
\name{as.character,CFtime-method}
\alias{as.character,CFtime-method}
\title{Return the timestamps contained in the CFtime instance.}
\usage{
\S4method{as.character}{CFtime}(x)
}
\arguments{
\item{x}{The CFtime instance whose timestamps will be returned}
}
\value{
The timestamps in the specified CFtime instance.
}
\description{
Return the timestamps contained in the CFtime instance.
}
\examples{
cf <- CFtime("days since 1850-01-01", "julian", 0:364)
as.character(cf)
}
