% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{as.character.CFTime}
\alias{as.character.CFTime}
\title{Return the timestamps contained in the \code{CFTime} instance.}
\usage{
\method{as.character}{CFTime}(x, ...)
}
\arguments{
\item{x}{The \code{CFTime} instance whose timestamps will be returned.}

\item{...}{Ignored.}
}
\value{
The timestamps in the specified \code{CFTime} instance.
}
\description{
Return the timestamps contained in the \code{CFTime} instance.
}
\examples{
t <- CFtime("days since 1850-01-01", "julian", 0:364)
as.character(t)
}
