% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_timestamp.R
\name{CF_parse_timestamp}
\alias{CF_parse_timestamp}
\title{Parse series of timestamps in CF format to date-time elements}
\usage{
CF_parse_timestamp(x, calendar = "standard")
}
\arguments{
\item{x}{character. Vector of character string representing timestamps in
ISO8601 extended or UDUNITS broken format.}

\item{calendar}{character. Atomic character string indicating the CF calendar
to use in interpreting the validity of the date.}
}
\value{
A data frame with constituent elements of the parsed timestamps in
numeric format. The columns are year, month, day, hour, minute, second (with
an optional fraction), tz_hour and tz_minute.
Missing and invalid input data will appear as \code{NA} - if the year is \code{NA} then
the timestamp in its entirety is invalid, otherwise \code{NA} values specify
missing information for which defaults exist.
}
\description{
This function will parse a vector of timestamps in ISO8601 or UDUNITS format
into a data frame with columns for the elements of the timestamp: year, month,
day, hour, minute, second, time zone hour and minute. Those timestamps that
could not be parsed or which represent an invalid date in the indicated
calendar will have \code{NA} values for the elements of the offending timestamp.
}
\details{
The supported formats are the \emph{broken timestamp} format from the UDUNITS
library and ISO8601 \emph{extended}, both with minor changes, as suggested by the
CF Metadata Conventions 1.9. In general, the format is \verb{YYYY-MM-DD hh:mm:ss.sss hh:mm}.
The year can be from 1 to 4 digits and is interpreted literally, so \code{79-10-24}
is the day Mount Vesuvius erupted and destroyed Pompeii, not \code{1979-10-24}
(when \href{https://www.uefa.com/uefachampionsleague/match/63624--ajax-vs-omonia/}{Ajax Amsterdam beat Omonoia Cyprus 10-0 in the first leg of the second round of the European Cup competition, with 5 goals from Søren Lerby},
as everybody vividly remembers). The year and month are mandatory,
all other fields are optional. There are defaults for all missing values,
following the UDUNITS and CF Metadata Conventions. Leading zeros can be omitted in the
UDUNITS format. The optional fractional part can have as many digits as the
precision calls for and will be applied to the smallest specified time unit.
In the result of this function if the fraction is associated with the minute
or the hour, it is converted into a regular \code{hh:mm:ss.sss} format, i.e. any
fraction in the result is always associated with the second.
The time zone is optional and should have at least the hour or \code{Z} if present,
the minute is optional. The time zone hour can have an optional sign. The
separator between the date and the time can be a single whitespace character
or a \code{T}; in the UDUNITS format the separator between the time and the time
zone must be a single whitespace character.

Currently only the extended formats (with separators between the elements)
are supported. The vector of timestamps may have any combination of ISO8601
and UDUNITS formats.
}
\examples{
timestamps <- c("2012-01-01T12:21:34Z", "12-1-23", "today",
                "2022-08-16T11:07:34.45-10", "2022-08-16 10.5+04")
x <- CF_parse_timestamp(timestamps, "proleptic_gregorian")
}
