% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFdatum.R
\docType{class}
\name{CFdatum-class}
\alias{CFdatum-class}
\title{CFdatum class}
\value{
An object of class CFdatum
}
\description{
This class stores the information to represent date and time values using
the CF conventions.
}
\details{
The following calendars are supported:

\itemize{
\item \code{gregorian} or \code{standard}, the international standard calendar for civil use.
\item \code{proleptic_gregorian}, the common calendar but extending before 1582-10-15
when the Gregorian calendar was adopted.
\item \code{noleap} or \verb{365_day}, all years have 365 days.
\item \code{all_leap} or \verb{366_day}, all years have 366 days.
\item \verb{360_day}, all years have 360 days, divided over 12 months of 30 days.
\item \code{julian}, every fourth year is a leap year (so including the years 1700, 1800, 1900, 2100, etc).
}
}
\section{Slots}{

\describe{
\item{\code{definition}}{character. The string that defines the time unit and base date/time.}

\item{\code{unit}}{numeric. The unit of time in which offsets are expressed.}

\item{\code{origin}}{numeric. Vector of components year, month, day and seconds.}

\item{\code{tz}}{character. Time zone offset hour and minute.}

\item{\code{calendar}}{character. The CF-calendar for the instance.}

\item{\code{cal_id}}{numeric. The internal identifier of the CF-calendar to use.}
}}

