% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFdatum.R
\name{CFdatum}
\alias{CFdatum}
\title{Create a CFdatum object}
\usage{
CFdatum(definition, calendar)
}
\arguments{
\item{definition}{character. A character string describing the time coordinate
of a CF-compliant data file.}

\item{calendar}{character. A character string describing the calendar to use
with the time dimension definition string.}
}
\value{
An object of the \code{CFdatum} class.
}
\description{
This function creates an instance of the \code{CFdatum} class. After creation the
instance is read-only. The parameters to the call are typically read from a
CF-compliant data file with climatological observations or predictions.
}
