% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFtime.R
\name{CFrange}
\alias{CFrange}
\alias{CFrange,CFtime-method}
\title{Extreme time series values}
\usage{
CFrange(x, format = "")

\S4method{CFrange}{CFtime}(x, format = "")
}
\arguments{
\item{x}{An instance of the \code{CFtime} class}

\item{format}{Atomic character string with format specifiers, optional}
}
\value{
character. Vector of two character representations of the extremes of the time series.
}
\description{
Character representation of the extreme values in the time series
}
\section{Methods (by class)}{
\itemize{
\item \code{CFrange(CFtime)}: Extreme values of the time series

}}
\examples{
cf <- CFtime("days since 1850-01-01", "julian", 0:364)
CFrange(cf)
CFrange(cf, "\%Y-\%b-\%e")
}
