% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFts.R
\name{CFts}
\alias{CFts}
\title{Create a CFts object}
\usage{
CFts(datum, offsets)
}
\arguments{
\item{datum}{CFdatum. An atomic instance of the \code{CFdatum} class describing
the date-time system upon which to base the offsets.}

\item{offsets}{numeric. A vector of offsets from the origin in the \code{datum}
instance. The unit of measure is defined by the \code{datum} instance as well.}
}
\value{
An object of the CFts class.
}
\description{
This function creates an instance of the CFts class. After creation the
instance is read-only. The parameters to the call are typically read from a
CF-compliant data file with climatological observations or predictions.
}
\examples{
datum <- CFdatum("days since 1850-01-01", "julian")
ts <- CFts(datum, 0:364)
}
