% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFdatum.R
\name{==,CFdatum,CFdatum-method}
\alias{==,CFdatum,CFdatum-method}
\title{Equivalence of CFdatum objects}
\usage{
\S4method{==}{CFdatum,CFdatum}(e1, e2)
}
\arguments{
\item{e1, e2}{CFdatum Instances of the CFdatum class.}
}
\value{
\code{TRUE} if the CFdatum objects are equivalent, \code{FALSE} otherwise.
}
\description{
This operator can be used to test if two CFdatum objects represent the same datum
for CF-convention time coordinates. Two CFdatum objects are considered equivalent
if they have the same definition string and the same calendar. Calendars
"standard", "gregorian" and "proleptic_gregorian" are considered equivalent,
as are the pairs of "365_day" and "no_leap", and "366_day" and "all_leap".
}
\examples{
e1 <- CFdatum("days since 1850-01-01", "gregorian")
e2 <- CFdatum("days since 1850-01-01 00:00:00", "standard")
e1 == e2
}
