% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFts.R
\name{==,CFts,CFts-method}
\alias{==,CFts,CFts-method}
\title{Equivalence of CFts objects}
\usage{
\S4method{==}{CFts,CFts}(e1, e2)
}
\arguments{
\item{e1, e2}{CFts. Instances of the CFts class.}
}
\value{
\code{TRUE} if the CFts objects are equivalent, \code{FALSE} otherwise.
}
\description{
This operator can be used to test if two CFts objects represent the same
CF-convention time coordinates. Two CFts objects are considered equivalent
if they have an equivalent datum and the same offsets.
}
\examples{
e1 <- CFts(CFdatum("days since 1850-01-01", "gregorian"), 0:364)
e2 <- CFts(CFdatum("days since 1850-01-01 00:00:00", "standard"), 0:364)
e1 == e2
}
