% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFtime.R
\name{length,CFtime-method}
\alias{length,CFtime-method}
\title{The length of the offsets contained in the CFtime instance.}
\usage{
\S4method{length}{CFtime}(x)
}
\arguments{
\item{x}{The CFtime instance whose length will be returned}
}
\value{
The number of offsets in the specified CFtime instance.
}
\description{
The length of the offsets contained in the CFtime instance.
}
\examples{
cf <- CFtime("days since 1850-01-01", "julian", 0:364)
length(cf)
}
