% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{length.CFTime}
\alias{length.CFTime}
\title{The length of the offsets contained in the \code{CFTime} instance.}
\usage{
\method{length}{CFTime}(x)
}
\arguments{
\item{x}{The \code{CFTime} instance whose length will be returned}
}
\value{
The number of offsets in the specified \code{CFTime} instance.
}
\description{
The length of the offsets contained in the \code{CFTime} instance.
}
\examples{
t <- CFtime("days since 1850-01-01", "julian", 0:364)
length(t)
}
