% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_days.R
\name{month_days}
\alias{month_days}
\title{Return the number of days in a month given a certain CF calendar}
\usage{
month_days(x, datum)
}
\arguments{
\item{x}{A vector of dates as strings with format \code{YYYY-MM-DD}.}

\item{datum}{The CFdatum specification to use.}
}
\value{
A vector indicating the number of days in each month for the vector of
dates supplied as a parameter to the function
}
\description{
Given a vector of dates as strings in ISO 8601 format and a
CFdatum object, this function will return a vector of the same length as the
dates, indicating the number of days in the month according to the CFdatum
specification.
}
\examples{
dates <- c("2021-11-27", "2021-12-10", "2022-01-14", "2022-02-18")
datum <- CFdatum("days since 1850-01-01", "standard")
month_days(dates, datum)

datum <- CFdatum("days since 1850-01-01", "all_leap")
month_days(dates, datum)

datum <- CFdatum("days since 1850-01-01", "360_day")
month_days(dates, datum)

}
