% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFts.R
\name{+,CFts,numeric-method}
\alias{+,CFts,numeric-method}
\title{Extend an CFts object with additional offsets}
\usage{
\S4method{+}{CFts,numeric}(e1, e2)
}
\arguments{
\item{e1}{CFts. Instance of the CFts class.}

\item{e2}{numeric. Vector of offsets to be added to the CFts instance.}
}
\value{
A CFts object with offsets composed of the CFts instance and the
numeric vector.
}
\description{
A CFts instance can be extended by adding additional offsets using this
operator.
}
\details{
The resulting CFts instance will have its offsets in the order that they are added,
meaning that the offsets from the CFts instance come first and those from the
numeric vector follow. There is no reordering or removal of duplicates. This
is because the time series are usually associated with a dataset and the
correspondence between the two is thus preserved.

Note that when adding multiple vectors of offsets to a CFts instance, it is
more efficient to first concatenate the vectors and then do a final addition
to the CFts instance. So avoid \code{e1 + e2 + e3 + ...} but rather do
\code{e1 + c(e2, e3, ...)}.
}
\examples{
e1 <- CFts(CFdatum("days since 1850-01-01", "gregorian"), 0:364)
e2 <- 365:729
e1 + e2
}
