% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{slab}
\alias{slab}
\title{Which time steps fall within two extreme values}
\usage{
slab(x, extremes, rightmost.closed = FALSE)
}
\arguments{
\item{x, extremes, rightmost.closed}{See \code{slice()}.}
}
\value{
See \code{slice()}.
}
\description{
Avoid using this function, use \code{\link[=slice]{slice()}} instead. This function will be
deprecated in the near future.
}
\examples{
t <- CFtime("hours since 2023-01-01 00:00:00", "standard", 0:23)
slab(t, c("2022-12-01", "2023-01-01 03:00"))
}
