% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFutils.R
\name{str.CFdatum}
\alias{str.CFdatum}
\title{Display the structure of a CFdatum instance}
\usage{
\method{str}{CFdatum}(object, ...)
}
\arguments{
\item{object}{\code{CFdatum} instance to print structure of.}

\item{...}{Ignored.}
}
\value{
Nothing. Prints information to the console.
}
\description{
Display the structure of a CFdatum instance
}
