% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFtime.R
\name{str.CFTime}
\alias{str.CFTime}
\title{Compact display of a CFTime instance}
\usage{
\method{str}{CFTime}(object, ...)
}
\arguments{
\item{object}{A \code{CFTime} instance.}

\item{...}{Ignored.}
}
\description{
Compact display of a CFTime instance
}
