% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFtime.R
\docType{class}
\name{CFtime-class}
\alias{CFtime-class}
\title{CF Metadata Conventions time representation}
\value{
An object of class CFtime.
}
\description{
CF Metadata Conventions time representation
}
\section{Slots}{

\describe{
\item{\code{datum}}{CFdatum. The atomic origin upon which the \code{offsets} are based.}

\item{\code{resolution}}{numeric. The average number of time units between offsets.}

\item{\code{time}}{data.frame. Data frame of numeric date elements year, month, day,
hour, minute, second and time zone from offsets defined by the \code{datum}, as
well as the offset itself.}
}}

