% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFtime.R
\docType{class}
\name{CFtime-class}
\alias{CFtime-class}
\title{CF Metadata Conventions time representation}
\value{
An object of class CFtime.
}
\description{
CF Metadata Conventions time representation
}
\section{Slots}{

\describe{
\item{\code{datum}}{CFdatum. The atomic origin upon which the \code{offsets} are based.}

\item{\code{resolution}}{numeric. The average number of time units between offsets.}

\item{\code{offsets}}{numeric. A vector of offsets from the datum.}
}}

