% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated_functions}
\alias{deprecated_functions}
\alias{CFtimestamp}
\alias{CFmonth_days}
\alias{CFcomplete}
\title{Deprecated functions}
\usage{
CFtimestamp(cf, format = NULL, asPOSIX = FALSE)

CFmonth_days(cf, x = NULL)

CFcomplete(x)
}
\arguments{
\item{cf, x, format, asPOSIX}{See replacement functions.}
}
\value{
See replacement functions.
}
\description{
These functions are deprecated and should no longer be used in new code. The
below table gives the replacement function to use instead. The function
arguments of the replacement function are the same as those of the deprecated
function if no arguments are given in the table.\tabular{ll}{
   \strong{Deprecated function} \tab \strong{Replacement function} \cr
   CFcomplete() \tab \code{\link[=is_complete]{is_complete()}} \cr
   CFmonth_days() \tab \code{\link[=month_days]{month_days()}} \cr
   CFrange() \tab \code{\link[=range]{range()}} \cr
   CFsubset() \tab \code{\link[=slab]{slab()}} \cr
   CFtimestamp() \tab \code{\link[=as_timestamp]{as_timestamp()}} \cr
}
}
