% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFts.R
\name{+,CFts,CFts-method}
\alias{+,CFts,CFts-method}
\title{Merge two CFts objects}
\usage{
\S4method{+}{CFts,CFts}(e1, e2)
}
\arguments{
\item{e1, e2}{CFts. Instances of the CFts class.}
}
\value{
A CFts object with a set of offsets equal to the offsets of the
instances of CFts that the operator operates on. If the datums of the CFts
instances are not equivalent, an error is thrown.
}
\description{
Two CFts instances can be merged into one with this operator, provided that
the datum of the two instances is equivalent.
}
\details{
The order of the two parameters is significant. The resulting CFts will have
the offsets of both instances in the order that they are specified. There is
no reordering or removal of duplicates. This is because the time series are
usually associated with a dataset and the correspondence between the two is
thus preserved.
}
\examples{
e1 <- CFts(CFdatum("days since 1850-01-01", "gregorian"), 0:364)
e2 <- CFts(CFdatum("days since 1850-01-01 00:00:00", "standard"), 0:364)
e1 + e2
}
